trying out biseq instead of bstrcmp. redusing cxtracker hash to 32 bit

+++ src/cxt.c
@@ -19,7 +19,7 @@ int cx_track(struct in6_addr ip_src, uint16_t src_port,
 
     connection *cxt = NULL;
     connection *head = NULL;
-    uint64_t hash;
+    uint32_t hash;
 
     if (af == AF_INET) {
         hash = ((ip_src.s6_addr32[0] + ip_dst.s6_addr32[0])) % BUCKET_SIZE;
@@ -132,41 +132,13 @@ int cx_track(struct in6_addr ip_src, uint16_t src_port,
         cxt->d_total_pkts = 0;
         cxt->start_time = tstamp;
         cxt->last_pkt_time = tstamp;
-
         cxt->s_ip = ip_src;
         cxt->d_ip = ip_dst;
-
-        /*
-         * if (af = AF_INET) { 
-         */
-        /*
-         * cxt->s_ip6.s6_addr32[1]          = 0; 
-         */
-        /*
-         * cxt->s_ip6.s6_addr32[2]          = 0; 
-         */
-        /*
-         * cxt->s_ip6.s6_addr32[3]          = 0; 
-         */
-        /*
-         * cxt->d_ip6.s6_addr32[1]          = 0; 
-         */
-        /*
-         * cxt->d_ip6.s6_addr32[2]          = 0; 
-         */
-        /*
-         * cxt->d_ip6.s6_addr32[3]          = 0; 
-         */
-        /*
-         * } 
-         */
-
         cxt->s_port = src_port;
         cxt->d_port = dst_port;
         cxt->proto = ip_proto;
         cxt->next = head;
         cxt->prev = NULL;
-
         /*
          * New connections are pushed on to the head of bucket[s_hash] 
          */
@@ -175,7 +147,7 @@ int cx_track(struct in6_addr ip_src, uint16_t src_port,
         /*
          * Return value should be 1, telling to do client service fingerprinting 
          */
-        return 1;               // Client - check!
+        return 1;
     }
     /*
      * Should never be here!
